requirement.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Cython==0.29.27
dill==0.3.4
dm-tree==0.1.6
fastapi==0.73.0
filelock==3.4.2
flatbuffers==2.0
fonttools==4.29.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
h11==0.13.0
h5py==3.6.0
httplib2==0.20.2
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
joblib==1.1.0
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.1
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
pluggy==1.0.0
portalocker==2.3.2
promise==2.3
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydantic==1.9.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacrebleu==2.0.0
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-datasets==4.5.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-metadata==1.6.0
tensorflow-model-optimization==0.7.0
tensorflow-text==2.7.3
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.7.0
tf-slim==1.1.0
threadpoolctl==3.0.0
tokenizers==0.11.4
toml==0.10.2
tqdm==4.62.3
transformers==4.16.1
typeguard==2.13.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
